# patch rexml to improve performance of multi-threaded xml parsing