--- a/.travis.yml
+++ b/.travis.yml
@@ -5,5 +5,5 @@ import:
 env:
   jobs:
     # lock on version 8.x because use Ruby 2.6.0 available from 8.4.0
-    - ELASTIC_STACK_VERSION=8.x DOCKER_ENV=dockerjdk17.env
-    - SNAPSHOT=true ELASTIC_STACK_VERSION=8.x DOCKER_ENV=dockerjdk17.env
+    - ELASTIC_STACK_VERSION=8.x DOCKER_ENV=dockerjdk21.env
+    - SNAPSHOT=true ELASTIC_STACK_VERSION=8.x DOCKER_ENV=dockerjdk21.env