This change monkey patches the `SourceFactory::create_from` method from rexml to not require 'stringio'. This is done by copying that method to this plugin and moving the require call to the top.

To check the performance improvement. Before:

```
❯ time bin/logstash -e 'input { java_generator { message => "<root> <item>&#x1111;&#x1111;&#x1111;&#x1111;&#x1111;&#x1111;&#x1111;&#x1111;&#x1111;&#x1111;</item> <item>&#x2222;&#x2222;&#x2222;&#x2222;&#x2222;&#x2222;&#x2222;&#x2222;&#x2222;&#x2222;</item> <item>&#x3333;&#x3333;&#x3333;&#x3333;&#x3333;&#x3333;&#x3333;&#x3333;&#x3333;&#x3333;</item> <item>&#x4444;&#x4444;&#x4444;&#x4444;&#x4444;&#x4444;&#x4444;&#x4444;&#x4444;&#x4444;</item> <item>&#x5555;&#x5555;&#x5555;&#x5555;&#x5555;&#x5555;&#x5555;&#x5555;&#x5555;&#x5555;</item> <item>&#x6666;&#x6666;&#x6666;&#x6666;&#x6666;&#x6666;&#x6666;&#x6666;&#x6666;&#x6666;</item> </root>" count => 100000 } } filter { xml { target => target source => message } } output { null {} }' --log.level=error
Using bundled JDK: /tmp/test/logstash-8.15.3.slow/jdk.app/Contents/Home
Sending Logstash logs to /tmp/test/logstash-8.15.3.slow/logs which is now configured via log4j2.properties
bin/logstash -e  --log.level=error  100.31s user 25.54s system 242% cpu 51.794 total
```

After:

```
❯ time bin/logstash -e 'input { java_generator { message => "<root> <item>&#x1111;&#x1111;&#x1111;&#x1111;&#x1111;&#x1111;&#x1111;&#x1111;&#x1111;&#x1111;</item> <item>&#x2222;&#x2222;&#x2222;&#x2222;&#x2222;&#x2222;&#x2222;&#x2222;&#x2222;&#x2222;</item> <item>&#x3333;&#x3333;&#x3333;&#x3333;&#x3333;&#x3333;&#x3333;&#x3333;&#x3333;&#x3333;</item> <item>&#x4444;&#x4444;&#x4444;&#x4444;&#x4444;&#x4444;&#x4444;&#x4444;&#x4444;&#x4444;</item> <item>&#x5555;&#x5555;&#x5555;&#x5555;&#x5555;&#x5555;&#x5555;&#x5555;&#x5555;&#x5555;</item> <item>&#x6666;&#x6666;&#x6666;&#x6666;&#x6666;&#x6666;&#x6666;&#x6666;&#x6666;&#x6666;</item> </root>" count => 100000 } } filter { xml { target => target source => message } } output { null {} }' --log.level=error
Using bundled JDK: /tmp/test/logstash-8.15.3.fast/jdk.app/Contents/Home
Sending Logstash logs to /tmp/test/logstash-8.15.3.fast/logs which is now configured via log4j2.properties
bin/logstash -e  --log.level=error  121.21s user 2.96s system 682% cpu 18.194 total
```

closes #83