@yaauie 2024-10-29
Is this worth upstreaming to [ruby/rexml](https://github.com/ruby/rexml/issues)?