@jsvd 2024-10-30
@yaauie I have submitted a PR upstream but I don't have a lot of hope there since it's a JRuby specific thing (maybe less so now with Ractors in Ruby 3.x) https://github.com/ruby/rexml/pull/219